refactor : Product.java
- 상품 클래스에서 쓰이는 메서드들 이동
